Parallelise S3 upload tasks
S3 is being a bit too slow for my liking without this.

Also includes not re-initialising the S3 client on every request which I
am starting to suspect is just wrong. It's what spectator server does,
but doing that causes the 32-client pool to be recreated every time
which... given .NET guidelines on `HttpClient` usage, does... not seem
optimal...