Add user-facing validation for DB-enforced length limits of string properties
Closes https://github.com/ppy/osu-server-beatmap-submission/issues/37.

Similarly to
https://github.com/ppy/osu-server-beatmap-submission/pull/36, this won't
make the submission not-fail, but errors won't get reported to sentry
and users will get better feedback about why their map won't upload.

This also moves the romanised validations from
https://github.com/ppy/osu-server-beatmap-submission/pull/24 to the
built-in data annotations machinery as it's honestly just easier to not
reinvent the wheel, and the declarativeness of the syntax is also nice.