Fix validation attributes crashing on overflow
Closes https://github.com/ppy/osu-server-beatmap-submission/issues/42.

This is a bit of a dumb one unfortunately...

`RangeAttribute` has 2 constructors that accept number literals; one
takes `int`s and one takes `double`s. The actual range comparison
later on uses a bunch of type conversion magic and such, and the
conversion takes place either using `int` or `double`. In this case,
quantities out of range of `int` but *in* range of `uint` were failing
on attempts to convert them to `int` before even attempting to compare
them to the range bounds specified.

There's a third constructor of `RangeAttribute` that accepts a type and
two... strings (okay, that's probably an attribute limitation, but
*weird*). Which should not die in this way, so we use it.