Show better error message on exceeding absolute request body cap
This got dinged a few times yesterday on sentry
(https://sentry.ppy.sh/organizations/ppy/issues/73418/?project=12)
because somebody hit the absolute 200MB cap a few times.

This change is designed to (a) stop that from dinging on sentry because
there's really no reason to log it as error, and (b) give the user a
better error message to work with because the current one was kinda
opaque (would just show "BadRequest" or similar).