Fix incorrect validation of circle size / key count
The big issue is that the range specified on the attribute was [1, 18]
which only makes sense in mania.

The small issue that while the range was checked, nothing decimal should
really be allowed for mania anyway.

A significant amount of this is kind of much-of-muchness because
`LegacyBeatmapDecoder` already does clamping on this:

https://github.com/ppy/osu/blob/834d63d1deacf0be0c4e76cd5d47cee136fbb49a/osu.Game/Beatmaps/Formats/LegacyBeatmapDecoder.cs#L113-L131

so the checks are kind of dead in practice, but I'd rather have them
not, what with users' propensity to try and break stuff and not have us
notice that it happened.

The game package bump here is significant because it exposes
`LegacyBeatmapDecoder.MAX_MANIA_KEY_COUNT`.