Fix calculation of `total_length` being inconsistent with old submission
Compare
https://github.com/peppy/osu-web-10/blob/18ae5b4f9dc6e9637d69e47460e6affb5840d3ef/www/web/osu-osz2-bmsubmit-upload.php#L151.

Note that `Math.Ceiling()` is chosen intentionally rather than
`Math.Floor()`. The reason for this is that web-10 flooring is
technically incorrect if we are to assume editor timestamps are
validated against `total_length` because you could have a timestamp
that's just a couple of milliseconds past the
floored-time-of-last-object.

Reported in
https://discord.com/channels/90072389919997952/1259818301517725707/1436493355520626690.